Move struct ForwardedPort to library
